requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
gunicorn>=21.2
psycopg2-binary>=2.9
Pillow>=10.0
pyotp>=2.9
qrcode>=7.4
